change back to curl...

+++ utils/user_id.cpp
@@ -2,7 +2,7 @@
 
 int64_t get_userid(const std::wstring &s){
     int64_t ans = 0;
-    for(int i = 0; i < s.length(); i++){
+    for(size_t i = 0; i < s.length(); i++){
         if(L'0'<=s[i] && s[i]<=L'9'){
             ans = ans * 10 + s[i] - L'0';
         }
@@ -11,7 +11,7 @@ int64_t get_userid(const std::wstring &s){
 }
 int64_t get_userid(const std::string &s){
     int64_t ans = 0;
-    for(int i = 0; i < s.length(); i++){
+    for(size_t i = 0; i < s.length(); i++){
         if('0'<=s[i] && s[i]<='9'){
             ans = ans * 10 + s[i] - '0';
         }
